Make ParsedPort public
